requirements.txt
fastapi
uvicorn
ltp
pydantic
fire
httpx
PyYAML